move getIsBrowserPrerenderBroken into browser-runtime.utils
