+++ app/scripts/contentscript.js
@@ -5,10 +5,9 @@ import pump from 'pump';
 import { obj as createThoughStream } from 'through2';
 import browser from 'webextension-polyfill';
 import { EXTENSION_MESSAGES } from '../../shared/constants/app';
-import { checkForLastError } from '../../shared/modules/browser-runtime.utils';
+import { checkForLastError, getIsBrowserPrerenderBroken } from '../../shared/modules/browser-runtime.utils';
 import { isManifestV3 } from '../../shared/modules/mv3.utils';
 import shouldInjectProvider from '../../shared/modules/provider-injection';
-import { getIsBrowserPrerenderBroken } from '../../ui/helpers/utils/util';
 
 // contexts
 const CONTENT_SCRIPT = 'metamask-contentscript';
